Move extension & file loading out of QuPathGUI
Revise and simplify code for ExtensionClassLoader, making more use of Path than File
