+++ qupath-gui-fx/src/main/java/qupath/lib/gui/prefs/QuPathStyleManager.java
@@ -23,18 +23,21 @@
 
 package qupath.lib.gui.prefs;
 
+import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.Method;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardWatchEventKinds;
 import java.nio.file.WatchEvent;
 import java.nio.file.WatchKey;
 import java.nio.file.WatchService;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
@@ -52,6 +55,9 @@ import javafx.collections.ListChangeListener.Change;
 import javafx.collections.ObservableList;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.common.ThreadTools;
+import qupath.lib.gui.commands.Commands;
+import qupath.lib.gui.dialogs.Dialogs;
+import qupath.lib.gui.dialogs.Dialogs.DialogButton;
 import qupath.lib.gui.tools.GuiTools;
 
 
@@ -249,6 +255,74 @@ public class QuPathStyleManager {
 				selectedStyle.set(previouslySelected);
 		}
 	}
+	
+	
+	/**
+     * Handle installing CSS files (which can be used to style QuPath).
+     * @param list list of css files
+     * @return
+     */
+	public static boolean installStyles(final Collection<File> list) {
+		var dir = Commands.requestUserDirectory(true);
+		if (dir == null)
+			return false;
+		
+		var pathCssString = PathPrefs.getCssStylesPath();
+		
+		int nInstalled = 0;
+		try {
+			// If we have a user directory, add a CSS subdirectory if needed
+			var pathCss = Paths.get(pathCssString);
+			if (!Files.exists(pathCss)) {
+				if (Files.isDirectory(pathCss.getParent()))
+					Files.createDirectory(pathCss);
+			}
+			// If we still don't have a css directory, return
+			if (!Files.isDirectory(pathCss))
+				return false;
+			
+			// Copy over the files
+			Boolean overwriteExisting = null;
+			for (var file : list) {
+				if (!file.getName().toLowerCase().endsWith(".css")) {
+					logger.warn("Cannot install style for {} - not a .css file!", file);
+					continue;
+				}
+				var source = file.toPath();
+				var target = pathCss.resolve(file.getName());
+				if (Objects.equals(source, target)) {
+					logger.warn("Can't copy CSS - source and target files are the same!");
+					continue;
+				}
+				if (Files.exists(target)) {
+					// Check if we want to overwrite - if so, retain the response so we don't 
+					// have to prompt multiple times if there are multiple files
+					if (overwriteExisting == null) {
+						var response = Dialogs.showYesNoCancelDialog("Install CSS", "Do you want to overwrite existing CSS files?");
+						if (response == DialogButton.YES)
+							overwriteExisting = Boolean.TRUE;
+						else if (response == DialogButton.NO)
+							overwriteExisting = Boolean.FALSE;
+						else // cancelled
+							return false;
+					}
+					// Skip
+					if (!overwriteExisting)
+						continue;
+				}
+				logger.info("Copying {} -> {}", source, target);
+				Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);	
+				nInstalled++;
+			}
+		} catch (IOException e) {
+			logger.error("Exception installing CSS files: " + e.getLocalizedMessage(), e);
+			return false;
+		}
+		if (nInstalled > 0)
+			QuPathStyleManager.updateAvailableStyles();
+		return true;
+	}
+	
 	
 	
 	/**
